First raw implementation of the project. TODO: Refactor, optimize, classify
