--- src/main/java/com/surfwear/graphql/GraphqlApplication.java
@@ -1,13 +1,0 @@
-package com.surfwear.graphql;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class GraphqlApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(GraphqlApplication.class, args);
-	}
-
-}
