--- src/main/java/com/surfwear/graphql/entities/DetailCommande.java
@@ -1,10 +1,0 @@
-package com.surfwear.graphql.entities;
-
-public class DetailCommande {
-
-    private int commandeId;
-    private int ArticleId;
-    private int qte;
-    private double prixUnitaire;
-
-}
