+++ src/main/java/com/uppa/surfwear/entities/Categorie.java
@@ -1,0 +1,46 @@
+package com.uppa.surfwear.entities;
+
+public class Categorie {
+    private int categorieId;
+    private String nomCategorie;
+    private String description;
+
+    public Categorie(int categorieId, String nomCategorie, String description) {
+        this.categorieId = categorieId;
+        this.nomCategorie = nomCategorie;
+        this.description = description;
+    }
+
+    public int getCategorieId() {
+        return categorieId;
+    }
+
+    public void setCategorieId(int categorieId) {
+        this.categorieId = categorieId;
+    }
+
+    public String getNomCategorie() {
+        return nomCategorie;
+    }
+
+    public void setNomCategorie(String nomCategorie) {
+        this.nomCategorie = nomCategorie;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    @Override
+    public String toString() {
+        return "Categorie{" +
+                "categorieId=" + categorieId +
+                ", nomCategorie='" + nomCategorie + '\'' +
+                ", description='" + description + '\'' +
+                '}';
+    }
+}
