+++ src/main/java/com/uppa/surfwear/entities/DetailCommande.java
@@ -1,0 +1,10 @@
+package com.uppa.surfwear.entities;
+
+public class DetailCommande {
+
+    private int commandeId;
+    private int ArticleId;
+    private int qte;
+    private double prixUnitaire;
+
+}
