+++ src/main/java/com/uppa/surfwear/interfaces/graphql/ISurfwearGraphqlController.java
@@ -1,0 +1,4 @@
+package com.uppa.surfwear.interfaces.graphql;
+
+public interface ISurfwearGraphqlController {
+}
