+++ src/main/java/com/uppa/surfwear/interfaces/graphql/ISurfwearGraphqlService.java
@@ -1,0 +1,4 @@
+package com.uppa.surfwear.interfaces.graphql;
+
+public interface ISurfwearGraphqlService {
+}
